refactor(ui): carve out app/state/_shared.ts as the cross-slice home
Scaffolding step for the per-domain slice work that follows. Pulls
the storage-format types, the queued-chat-messages storage key,
and the localStorage parse / normalize / serialize helpers out of
useRuntimeConsole.ts and into a new app/state/_shared.ts file.
No behaviour change — every helper round-trips byte-for-byte, the
imports are wired to the same callsites, and useRuntimeConsole.ts
keeps the same external shape.

What moves:
- ChatTarget, HecateChatTarget, QueuedChatMessage types
- queuedChatMessagesStorageKey constant
- normalizeStoredChatTarget (coercive — for non-storage callers
  where the wider type system has already vouched for the input)
- parseStoredChatTarget (strict — wipes corrupt localStorage keys
  per the usePersistedState contract)
- normalizeStoredHecateChatTarget (the Hecate-Chat subset)
- parseQueuedChatMessageList (per-item filter + array-shape
  rejection)
- parseChatTargetsBySessionID + serializeChatTargetsBySessionID
  (JSON object ↔ Map round-trip)

What stays in useRuntimeConsole.ts: the session-derive helpers
(agentChatSession{IsExternal,IsBusy},
deriveHecateChatTarget/SelectionFromSession). Those are
agentChats-domain specific and will move with the agentChats
slice when it's carved out.

This is the leaf-first confidence-builder before the four
slice extractions. Reviewers can verify the move-only diff has
no logic delta; the per-domain hooks that follow each get to
focus on their own state machine rather than re-litigating the
storage-format contract.
